requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Config
pyyaml>=6.0

# Model persistence
joblib>=1.3.0
